+++ backend/src/main/java/project/backend/auth/dto/OAuthMemberDto.java
@@ -1,0 +1,9 @@
+package project.backend.auth.dto;
+
+public record OAuthMemberDto(
+	String email,
+	String nickname,
+	String login
+) {
+
+}
