+++ backend/src/main/java/project/backend/auth/token/jwt/Token.java
@@ -1,0 +1,8 @@
+package project.backend.auth.token.jwt;
+
+public record Token(
+	String accessToken,
+	String refreshToken
+) {
+
+}
