+++ backend/src/main/java/project/backend/auth/token/jwt/TokenStatus.java
@@ -1,0 +1,9 @@
+package project.backend.auth.token.jwt;
+
+public enum TokenStatus {
+	VALID,
+	EXPIRED,
+	INVALID_SIGNATURE,
+	MALFORMED,
+	UNKNOWN_ERROR
+}
