--- backend/src/main/java/project/backend/global/common/ScrollPaginationCollection.java
@@ -1,34 +1,0 @@
-package project.backend.global.common;
-
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public class ScrollPaginationCollection<T> {
-
-	private final List<T> itemsWithNextCursor;
-	private final int size;
-
-	public static <T> ScrollPaginationCollection<T> of(List<T> itemsWithNextCursor, int size) {
-		return new ScrollPaginationCollection<>(itemsWithNextCursor, size);
-	}
-
-	public List<T> getCurrentScrollItems() {
-		if (isLastScroll()) {
-			return itemsWithNextCursor;
-		}
-		return itemsWithNextCursor.subList(0, size);
-	}
-
-	public T getNextCursor() {
-		if (isLastScroll()) {
-			return null;
-		}
-		return itemsWithNextCursor.get(size - 1);
-	}
-
-	public boolean isLastScroll() {
-		return itemsWithNextCursor.size() < size;
-	}
-}
