+++ backend/src/main/java/project/backend/global/security/config/SecurityConfig.java
@@ -11,15 +11,14 @@ import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.ExceptionTranslationFilter;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import project.backend.global.security.app.CustomOAuth2UserService;
+import project.backend.auth.app.CustomOAuth2UserService;
 import project.backend.global.security.handler.CustomLogoutSuccessHandler;
-import project.backend.global.security.handler.FormFailureHandler;
-import project.backend.global.security.handler.FormSuccessHandler;
-import project.backend.global.security.handler.JwtAuthenticationFilter;
-import project.backend.global.security.handler.OAuth2FailureHandler;
-import project.backend.global.security.handler.OAuth2SuccessHandler;
-import project.backend.global.security.handler.RestAuthenticationEntryPoint;
+import project.backend.global.security.handler.form.FormFailureHandler;
+import project.backend.global.security.handler.form.FormSuccessHandler;
+import project.backend.global.security.filter.JwtAuthenticationFilter;
+import project.backend.global.security.handler.oauth.OAuth2FailureHandler;
+import project.backend.global.security.handler.oauth.OAuth2SuccessHandler;
+import project.backend.global.security.entrypoint.RestAuthenticationEntryPoint;
 
 @Configuration
 @RequiredArgsConstructor
