+++ backend/src/main/java/project/backend/global/security/config/webconfig/WebConfig.java
@@ -1,0 +1,27 @@
+package project.backend.global.security.config.webconfig;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+	@Override
+	public void addCorsMappings(CorsRegistry registry) {
+		registry.addMapping("/**") // 모든 경로 허용
+			.allowedOrigins("http://localhost:3000") // React dev server
+			.allowedMethods("*")
+			.allowedHeaders("*")
+			.allowCredentials(true);
+	}
+
+	@Override
+	public void addResourceHandlers(ResourceHandlerRegistry registry) {
+		registry
+			.addResourceHandler("images/**") // 클라이언트 요청 URL
+			.addResourceLocations("file:./images/");   // 실제 서버 폴더 경로
+	}
+}
+
