--- backend/src/main/java/project/backend/global/security/dto/OAuthMemberDto.java
@@ -1,9 +1,0 @@
-package project.backend.global.security.dto;
-
-public record OAuthMemberDto(
-	String email,
-	String nickname,
-	String login
-) {
-
-}
