--- backend/src/main/java/project/backend/global/security/jwt/Token.java
@@ -1,8 +1,0 @@
-package project.backend.global.security.jwt;
-
-public record Token(
-	String accessToken,
-	String refreshToken
-) {
-
-}
