--- backend/src/main/java/project/backend/global/security/jwt/TokenStatus.java
@@ -1,9 +1,0 @@
-package project.backend.global.security.jwt;
-
-public enum TokenStatus {
-	VALID,
-	EXPIRED,
-	INVALID_SIGNATURE,
-	MALFORMED,
-	UNKNOWN_ERROR
-}
